% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coords}
\alias{coords}
\title{Geographical coordinates of Kansas counties' centroids (Longitude/Latitude, 2000)}
\format{
\describe{
\item{Longitude}{Longitude coordinate of the county centroid.}
\item{Latitude}{Latitude coordinate of the county centroid.}
}
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html}
}
\usage{
coords
}
\description{
A dataset containing the geographical coordinates of Kansas counties' centroids.
}
\keyword{datasets}
