% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coords_xy}
\alias{coords_xy}
\title{Geographical coordinates of Kansas counties' centroids (X/Y, Web Mercator, 2000)}
\format{
\describe{
\item{X}{X coordinate of the county centroid.}
\item{Y}{Y coordinate of the county centroid.}
}
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html}
}
\usage{
coords_xy
}
\description{
A dataset containing the geographical coordinates of Kansas counties' centroids
in Web Mercator projection.
}
\keyword{datasets}
