% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distance_mtp}
\alias{distance_mtp}
\title{Great-circle distances between Hérault municipalities, France}
\format{
A \code{matrix} with 342 rows and 342 columns. Each element represents the
distance between two municipalities. Municipality IDs are used as row and
column names.
}
\source{
\url{https://www.data.gouv.fr/datasets/contours-communes-france-administrative-format-admin-express-avec-arrondissements/}
}
\usage{
distance_mtp
}
\description{
A dataset containing the great-circle distance (in kilometers) between 342
Hérault municipalities.
}
\keyword{internal}
