% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_binaries.R
\name{install_binaries}
\alias{install_binaries}
\title{Download, unzip, check permission and test the bioRgeo's executable binary
files}
\usage{
install_binaries(binpath = NULL, infomap_version = c("2.1.0", "2.6.0"))
}
\arguments{
\item{binpath}{a \code{character}}

\item{infomap_version}{a \code{character} vector indicating the Infomap
version(s) to install.}
}
\description{
This function downloads and unzips the 'bin' folder needed to run some
functions of bioRgeo. It also checks if the files have the permissions to
be executed as programs. It finally tests if the executable binary files
are running properly.
}
\note{
Only the Infomap version 2.1.0 and 2.6.0 are available for now.
}
\author{
Maxime Lenormand (\email{maxime.lenormand@inrae.fr})
}
