% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mass}
\alias{mass}
\title{Population and number of out- and in-commuters by US Kansas county in 2000}
\format{
A \code{data.frame} with 105 rows and 3 columns:
\describe{
\item{rownames}{County ID.}
\item{Population}{Number of inhabitants.}
\item{Out-commuters}{Number of out-commuters.}
\item{In-commuters}{Number of in-commuters.}
}
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/tables/2000/county-to-county-worker-flow-files/}
}
\usage{
mass
}
\description{
A dataset containing the number of inhabitants, in-commuters and
out-commuters for 105 US Kansas counties in 2000.
}
\keyword{datasets}
