% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mass_mtp}
\alias{mass_mtp}
\title{Population and number of out- and in-commuters in Hérault municipalities, France (2020)}
\format{
A \code{data.frame} with 342 rows and 4 columns:
\describe{
\item{rownames}{Municipality ID.}
\item{Population}{Number of inhabitants.}
\item{Out_Commuters}{Number of out-commuters.}
\item{In_Commuters}{Number of in-commuters.}
}
}
\source{
\url{https://www.insee.fr/fr/statistiques/7630376}
}
\usage{
mass_mtp
}
\description{
A dataset containing the number of inhabitants, in-commuters, and out-commuters
for 342 Hérault municipalities in 2020.
}
\keyword{internal}
