% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtp}
\alias{mtp}
\title{Spatial distribution of Hérault municipalities, France (2020)}
\format{
\describe{
\item{ID}{Municipality ID.}
\item{Longitude}{Longitude coordinate of the municipality centroid.}
\item{Latitude}{Latitude coordinate of the municipality centroid.}
\item{Area}{Surface area of the municipality (in square kilometers).}
\item{geometry}{Geometry of the municipality.}
}
}
\source{
\url{https://www.data.gouv.fr/datasets/contours-communes-france-administrative-format-admin-express-avec-arrondissements/}
}
\usage{
mtp
}
\description{
A dataset containing the geometry of 342 Hérault municipalities.
}
\keyword{internal}
