% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{od}
\alias{od}
\title{Origin-Destination commuting matrix for Kansas counties, USA (2000)}
\format{
A \code{matrix} with 105 rows and 105 columns. Each element represents the
number of commuters between two counties. County IDs are used as row and
column names.
}
\source{
<https://www2.census.gov/programs-surveys/decennial/tables/2000/
county-to-county-worker-flow-files/>
}
\usage{
od
}
\description{
A dataset containing the number of commuters between 105 Kansas counties in
2000.
}
\keyword{datasets}
