% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{od_mtp}
\alias{od_mtp}
\title{Origin-Destination commuting matrix for Hérault municipalities, France (2020)}
\format{
A \code{matrix} with 342 rows and 342 columns. Each element represents the
number of commuters between two municipalities. Municipality IDs are used as
row and column names.
}
\source{
\url{https://www.insee.fr/fr/statistiques/7630376}
}
\usage{
od_mtp
}
\description{
A dataset containing the number of commuters between 342 Hérault municipalities
in 2020.
}
\keyword{internal}
