% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spcounty}
\alias{spcounty}
\title{Spatial distribution of US counties in 2000}
\format{
A
\describe{
\item{ID}{County ID.}
\item{Longitude}{Longitude coordinate of the centroid of the county.}
\item{Latitude}{Latitude coordinate of the centroid of the county.}
\item{Area}{Surface area of the county (in square kilometer).}
\item{geometry}{Geometry of the county.}
}
}
\source{
\url{https://www.sciencebase.gov/catalog/item/4f4e4a2ee4b07f02db615738}
}
\usage{
spcounty
}
\description{
A dataset containing the geometry of 3,108 US counties.
}
\keyword{datasets}
