% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coords}
\alias{coords}
\title{Geographical coordinates of US Kansas counties' centroids in 2000
(Longitude/Latitude)}
\format{
\describe{
\item{Longitude}{Longitude coordinate of the centroid of the county.}
\item{Latitude}{Latitude coordinate of the centroid of the county.}
}
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html}
}
\usage{
coords
}
\description{
A dataset containing the geographical coordinates of US Kansas counties'
centroids in 2000 (Longitude/Latitude).
}
\keyword{datasets}
