% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coords_xy}
\alias{coords_xy}
\title{Geographical coordinates of US Kansas counties' centroids in 2000
(X/Y based on Web Mercator)}
\format{
\describe{
\item{X}{X coordinate of the centroid of the county.}
\item{Y}{Y coordinate of the centroid of the county.}
}
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html}
}
\usage{
coords_xy
}
\description{
A dataset containing the geographical coordinates of US Kansas counties'
centroids in 2000 (X/Y based on Web Mercator).
}
\keyword{datasets}
