% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distance}
\alias{distance}
\title{Great-circle distances between US counties}
\format{
A \code{matrix} with 3,108 rows and 3,108 columns. Each element of the
matrix represents the distance between two counties. County ID as rownames
and colnames.
}
\source{
\url{https://www.sciencebase.gov/catalog/item/4f4e4a2ee4b07f02db615738}
}
\usage{
distance
}
\description{
A dataset containing the great-circle distance (in kilometers) between 3,108
US counties.
}
\keyword{datasets}
