% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distance}
\alias{distance}
\title{Great-circle distances between US Kansas counties}
\format{
A \code{matrix} with 105 rows and 105 columns. Each element of the
matrix represents the distance between two counties. County ID as rownames
and colnames.
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html}
}
\usage{
distance
}
\description{
A dataset containing the great-circle distance (in kilometer) between 105
US Kansas counties.
}
\keyword{datasets}
