% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_opportunities.R
\name{extract_opportunities}
\alias{extract_opportunities}
\title{Compute the number of opportunities between pairs of locations}
\usage{
extract_opportunities(opportunity, distance, check_names = FALSE)
}
\arguments{
\item{opportunity}{a numeric vector the number of opportunites per
location. The value should be positive.}

\item{distance}{a squared matrix representing the distance between locations.}

\item{check_names}{a boolean indicating if the ID location are used as vector
names, matrix rownames and colnames and if they should be checked
(see Details).}
}
\value{
A squared matrix in which each element represents the number of opportunities
between a pair of locations.
}
\description{
This function compute the number of opportunities between pairs of locations
as defined in \insertCite{Lenormand2016}{TDLM}.
For a given pair of location the number of opportunities between the location
of origin and the location of destination is based on the number of
opportunities in a circle of radius distance between origin and destination
centered in the origin (excluding the origin and destination).
}
\details{
\code{opportunity} and \code{distance} should be based on the same number of
locations sorted in the same order. It is recommended to use the location ID
as vector names, matrix rownames and matrix colnames and to set
\code{check_names = TRUE} to check that everything is ok (\code{check_names = FALSE} by
default).
}
\examples{
data(mass)
data(distance)
# sij=extract_opportunities(opportunity=as.numeric(mass[,1]),
# distance=distance,
# check_names=FALSE)

}
\references{
\insertRef{Lenormand2016}{TDLM}
}
\author{
Maxime Lenormand (\email{maxime.lenormand@inrae.fr})
}
