% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spatial_information.R
\name{extract_spatial_information}
\alias{extract_spatial_information}
\title{Extract the distance and average surface area from a spatial object}
\usage{
extract_spatial_information(geometry, names = NULL, stdist = FALSE)
}
\arguments{
\item{geometry}{a spatial object that can be handled by the \code{sf} package.}

\item{names}{.}

\item{stdist}{a boolean indicating if the ... (see Details).}
}
\value{
A list composed of two elements. The first element is a squared matrix
representing the great-circle distance (in kilometer) between locations. The
second element is the average surface area of the locations
(in squared kilometer).
}
\description{
This function returns a matrix of distances (in kilometer) along with the
average surface area of the locations (in squared kilometer).
}
\details{
the \link[sf:geos_measures]{st_distance} functions from
the \href{https://cran.rstudio.com/web/packages/sf/index.html}{sf} package.

#' the \link[sf:geos_measures]{st_area} functions from
the \href{https://cran.rstudio.com/web/packages/sf/index.html}{sf} package.
}
\note{
By default
}
\examples{
data(county)

}
\seealso{
\code{\link[=extract_opportunities]{extract_opportunities()}} \code{\link[=check_format_names]{check_format_names()}}
}
\author{
Maxime Lenormand (\email{maxime.lenormand@inrae.fr})
}
