% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{od}
\alias{od}
\title{Origin-Destination commuting matrix between US Kansas counties in 2000}
\format{
A \code{matrix} with 105 rows and 105 columns. Each element of the
matrix represents the number of commuters between two counties. County ID as
rownames and colnames.
}
\source{
\url{https://www2.census.gov/programs-surveys/decennial/tables/2000/county-to-county-worker-flow-files/}
}
\usage{
od
}
\description{
A dataset containing the number of commuters between 105 US Kansas counties
in 2000.
}
\keyword{datasets}
